+++ example/.sword/shim/process.js
@@ -1,0 +1,4 @@
+
+  // process shim
+  process.argv = ["/usr/local/bin/node","/Users/seho/Desktop/openSource/sword-function-framework/packages/cli/lib/index.cjs","dev"]
+