+++ example/src/index.ts
@@ -1,11 +1,11 @@
-import { useApp, usePipeline, usePlugin } from '@sword-code-practice/sword-framework';
+import { useApp, usePipeline, usePlugin } from '@swordjs/sword-framework';
 
 const plugin = usePlugin();
 const pipeline = usePipeline();
 
 const init = async () => {
   const app = await useApp();
-  app.implementApi();
+  await app.implementApi();
   // 启动服务器
   app.server.start();
 };