+++ packages/runtime/src/core/api.ts
@@ -17,7 +17,7 @@ import {
   apiResponseHeaders as apiServerResponseHeaders
 } from './platform/server/api';
 import { adaptUnicloudEvent } from './platform/unicloud/api';
-import type * as H3 from '@sword-code-practice/h3';
+import type * as H3 from '@swordjs/h3';
 import type { HttpContext, HttpInstructMethod, UnPromisify, Event, CustomHandlerReturn, HttpApiStatusResponse } from '../../../../typings/index';
 import type { ValidateProto } from './validate';
 import type { InterruptPipelineResult } from './pipeline';