+++ packages/runtime/src/core/schedule.ts
@@ -3,10 +3,10 @@ import type { CommandConfig } from '../../../../typings/config';
 
 // 支持异步加载依赖的列表
 const asyncDependencyList = [
-  '@sword-code-practice/h3',
-  '@sword-code-practice/sword-plugin-log',
-  '@sword-code-practice/sword-plugin-server',
-  '@sword-code-practice/sword-plugin-unicloud-log'
+  '@swordjs/h3',
+  '@swordjs/sword-plugin-log',
+  '@swordjs/sword-plugin-server',
+  '@swordjs/sword-plugin-unicloud-log'
 ] as const;
 
 // 配置一个缓存对象
@@ -17,19 +17,19 @@ const asyncDependencyCache: Record<
     module: any;
   }
 > = {
-  '@sword-code-practice/h3': {
+  '@swordjs/h3': {
     platforms: ['server'],
     module: undefined
   },
-  '@sword-code-practice/sword-plugin-log': {
+  '@swordjs/sword-plugin-log': {
     platforms: ['server'],
     module: undefined
   },
-  '@sword-code-practice/sword-plugin-server': {
+  '@swordjs/sword-plugin-server': {
     platforms: ['server'],
     module: undefined
   },
-  '@sword-code-practice/sword-plugin-unicloud-log': {
+  '@swordjs/sword-plugin-unicloud-log': {
     platforms: ['unicloud'],
     module: undefined
   }