+++ typings/index.d.ts
@@ -2,7 +2,7 @@ import './pipeline';
 import './hook';
 import type { LogType } from './log';
 import type { UnicloudEvent } from './unicloud';
-import type * as H3 from '@sword-code-practice/h3';
+import type * as H3 from '@swordjs/h3';
 
 export type UnPromisify<T> = T extends Promise<infer U> ? U : T;